move use cache to the activity page
